Adding first cut at a rule parser
In order to do user based blocking, the admin will have to define what
part of the request the user value is contained in. The current function
takes a string a identifies the part of the request (header or body),
the location (cookie, query string, tag, etc) and the field.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>