Finish moving configuration to main conf
The previous commit only moved the section of the config the directives
were taken from. This commit finishes the move by actually transferring
the configuration all the way to the main conf. In the future, these
values may also be available via the location conf but for now this
makes it easier to move forward on core functionality.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>