Recycle Redis connections
There was a pretty bad performance issue with establishing a connection
to Redis for every request. The overhead wasn't bad, but the TIME_WAIT
was killing performance. This commit recycles the connection with each
worker establishing their own connection as needed.

Additional work needs to be done here for reconnecting and connection
error state handling, but this proved to be a huge initial performance
boost.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>