Adding sample backend application
In order to test if the downstream headers are being set properly, there
needs to be a backing application to receive them. This commit adds a
very small sinatra application that returns the status of an actor if
the headers are set from the Repsheet module. An additional location
was added to nginx.conf to support this application.