Adding CIDR code.
This introduces CIDR blocks to librepsheet. The goal is that an IP
address should be able to be compared against a CIDR block in the case
of whitelisting or blacklisting. There are some improvements necessary
still on validation and the code needs to be wired in and exposed in the
library.