Fix header proccessing. Closes #16.
I misunderstood how NGINX processed this. The ngx_str_t data member
holds the entire cookie and the len member holds the length of the
string that should be consumed. I thought that the data member held only
the relevant portion of the string which led to this error. In the fix I
copied the the correct part of the string out of the data member and
passed that back to librepsheet. There is a test in place to make sure
this doesn't happen again.