Implements repsheet_connect and repsheet_reconnect.
This fixes the issue where Redis becomes slow to respond and the nature
of the request is time sensitive. The request will timeout with the
value specified in milliseconds and set an error on the Redis
context. Once that happens the connection is unusable and must be
reset. This is where repsheet_reconnect becomes useful.