Get rid of the use of KEYS.
This was never a good idea to begin with. It also might be causing a
pretty nasty performance issue. I wasn't able to get enough data to
confirm, but this is a pretty logical conclusion and an easy move to
make.

This does, however, introduce another breaking change for other clients
pushing data into the Repsheet Redis database. The insert operation for
a CIDR block will now have to do the following operations:

```
SET <block>:repsheet:cidr:<white|black>listed <reason>
SADD repsheet:cidr:<white|black>listed <block>
```

This makes the block a member of a set that can easily be retrieved. The
CIDR block lookup will still end up being O(n) for members of the set,
but no longer incurs the O(n) cost from the KEYS lookup (where n is the
number of keys in the Redis database). This gets rid of a possible two
linear scans through the keyspace.