WIP Staring a massive refactor
I started building a Redis module to handle the logic that librepsheet
previously handled. This refactoring represents both a move to use the
new Redis module as well as clean out old features that don't seem to
be in use or were poorly designed.

Note: this commit is not releasable because it does not provide the
reason for the action taken. This will be added in the next commit.